+++ watch/heuristics.py
@@ -234,17 +234,25 @@ IARPA_REAL_STATUS = {
 }
 
 
+"""
+For official color definitions see:
+
+    :func:`iarpa_smart_metrics.evaluation.Evaluation.get_sm_color` and
+    :func:`iarpa_smart_metrics.evaluation.Evaluation.get_gt_color`
+"""
 IARPA_CONFUSION_COLORS = {}
-IARPA_CONFUSION_COLORS['gt_true_neg'] = 'darkgreen'  # no IARPA color for this, make one up.
+IARPA_CONFUSION_COLORS['gt_true_neg'] = 'green'  # no IARPA color for this, make one up.
 IARPA_CONFUSION_COLORS['gt_true_pos'] = 'lime'
 IARPA_CONFUSION_COLORS['gt_false_pos'] = 'red'
 IARPA_CONFUSION_COLORS['gt_false_neg'] = 'black'
 IARPA_CONFUSION_COLORS['gt_positive_unbounded'] = "darkviolet"
 IARPA_CONFUSION_COLORS['gt_ignore'] = "lightsalmon"
 IARPA_CONFUSION_COLORS['gt_seen'] = "gray"
-IARPA_CONFUSION_COLORS['sm_pos_match'] = "orange"
-IARPA_CONFUSION_COLORS['sm_partially_wrong'] = "magenta"
-IARPA_CONFUSION_COLORS['sm_completely_wrong'] = "aquamarine"
+IARPA_CONFUSION_COLORS['sm_pos_match'] = "aquamarine"
+IARPA_CONFUSION_COLORS['sm_partially_wrong'] = "orange"
+IARPA_CONFUSION_COLORS['sm_completely_wrong'] = "magenta"
+
+IARPA_CONFUSION_COLORS['sm_ignore'] = "lightsalmon"  # no IARPA color for this, make one up
 
 
 def iarpa_assign_truth_confusion(truth_status, has_positive_match):
@@ -307,7 +315,9 @@ def iarpa_assign_pred_confusion(truth_match_statuses):
             pred_cfsn = 'sm_pos_match'
     elif 'gt_false_pos' in truth_cfsns:
         pred_cfsn = 'sm_completely_wrong'
-
+    else:
+        if set(truth_cfsns) == {'gt_ignore'}:
+            pred_cfsn = 'sm_ignore'
     return pred_cfsn
 
 
@@ -424,8 +434,8 @@ def hack_track_categories(track_catnames, task):
         >>> for kw in ub.named_product(basis):
         >>>     task = kw['task']
         >>>     track_catnames = kw['track_catnames']
+        >>>     kw['new_catnames'] = hack_track_categories(track_catnames, task)
         >>>     print('kw = {}'.format(ub.urepr(kw, nl=1)))
-        >>>     print(hack_track_categories(track_catnames, task))
 
     Example:
         >>> from watch.heuristics import *  # NOQA