+++ watch/tasks/depth/predict.py
@@ -17,7 +17,8 @@ from tqdm import tqdm
 
 from scipy import ndimage
 
-from . import modules_monkeypatch  # NOQA
+import geowatch_tpl  # NOQA
+
 from .datasets import WVRgbDataset, WVSuperRgbDataset
 # from .pl_highres_verify import MultiTaskModel, modify_bn, dfactor, local_utils
 from .pl_highres_verify import MultiTaskModel, modify_bn, dfactor